+++ Materials/AST.txt
@@ -1,0 +1,1 @@
+<2,0,(<4,0,(<36,0>),(<1,1>),(<6,0,(<3,0,(<36,0>),(<1,2>)>),(<3,0,(<36,0>),(<1,3>)>)>)>),(<4,0,(<36,0>),(<1,0>),(<6,0>),(<2,0,(<40,0,(<5,0,(<1,1>),(<6,0,(<0,1001>),(<0,143>)>)>)>),(<40,0,(<5,0,(<1,1>),(<6,0,(<0,150>),(<0,45>)>)>)>),(<40,0,(<5,0,(<1,1>),(<6,0,(<0,241>),(<0,653>)>)>)>),(<38,0,(<0,0>)>)>)>),(<4,0,(<36,0>),(<1,1>),(<6,0,(<3,0,(<36,0>),(<1,2>)>),(<3,0,(<36,0>),(<1,3>)>)>),(<2,0,(<35,0,(<8,0,(<1,2>),(<1,3>)>),(<2,0,(<34,0,(<20,0,(<1,2>),(<1,3>)>),(<18,0,(<1,2>),(<17,0,(<1,2>),(<1,3>)>)>),(<18,0,(<1,3>),(<17,0,(<1,3>),(<1,2>)>)>)>)>)>),(<38,0,(<1,2>)>)>)>)>
