+++ Materials/Log.html
@@ -6,6 +6,6 @@
 <font color = red>
 <font color = blue>
 Build started on: 0.000000
-Build   ended on: 0.015000
+Build   ended on: 0.000000
 
 ========== Build finished ==========
